+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 Used [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.31.0] - 2022-07-22
+### Added
+- Method `sbpCheck`
+- Method `sbpPayment`
+- Method `sbpClientCheck`
+- Method `sbpClientPayment`
+
 ## [v1.30.0] - 2022-06-17
 ### Changed
 - Added parameters filtering to method registerReceiptAsync